refactor: split main.py into separate modules for better organization
Break down monolithic main.py into focused modules:
- config.py: settings, validation, and NetBox API setup
- cli.py: CLI callbacks and signal handlers
- playbook.py: Ansible playbook templates and execution
- resources.py: resource file processing logic
- utils.py: utility functions for Git operations

AI-assisted: Claude Code
