+++ netbox_manager/config.py
@@ -1,0 +1,62 @@
+# SPDX-License-Identifier: Apache-2.0
+
+import sys
+import warnings
+
+from dynaconf import Dynaconf, Validator, ValidationError
+from loguru import logger
+import pynetbox
+import typer
+
+warnings.filterwarnings("ignore")
+
+settings = Dynaconf(
+    envvar_prefix="NETBOX_MANAGER",
+    settings_files=["settings.toml", ".secrets.toml"],
+    load_dotenv=True,
+)
+
+# NOTE: lazy validate configuration
+settings.validators.register(
+    Validator("DEVICETYPE_LIBRARY", is_type_of=str)
+    | Validator("DEVICETYPE_LIBRARY", is_type_of=None, default=None),
+    Validator("MODULETYPE_LIBRARY", is_type_of=str)
+    | Validator("MODULETYPE_LIBRARY", is_type_of=None, default=None),
+    Validator("RESOURCES", is_type_of=str)
+    | Validator("RESOURCES", is_type_of=None, default=None),
+    Validator("TOKEN", is_type_of=str),
+    Validator("URL", is_type_of=str),
+    Validator("IGNORE_SSL_ERRORS", is_type_of=bool)
+    | Validator(
+        "IGNORE_SSL_ERRORS",
+        is_type_of=str,
+        cast=lambda v: v.lower() in ["true", "yes"],
+        default=False,
+    ),
+    Validator("VERBOSE", is_type_of=bool)
+    | Validator(
+        "VERBOSE",
+        is_type_of=str,
+        cast=lambda v: v.lower() in ["true", "yes"],
+        default=False,
+    ),
+)
+
+try:
+    settings.validators.validate_all()
+except ValidationError as e:
+    logger.error(f"Error validating configuration: {e.details}")
+    raise typer.Exit()
+
+nb = pynetbox.api(settings.URL, token=settings.TOKEN)
+
+inventory = {
+    "all": {
+        "hosts": {
+            "localhost": {
+                "ansible_connection": "local",
+                "ansible_python_interpreter": sys.executable,
+            }
+        }
+    }
+}
